requirements.txt
browser-use
python-dotenv
PyYAML
httpx>=0.27.0
pydantic>=2.7.0
playwright>=1.45.0
tenacity>=8.2.3
colorama
fastapi>=0.112.0
twilio>=9.2.3
aiosmtplib>=3.0.1
